+++ magentoerpconnect/partner.py
@@ -21,8 +21,10 @@
 
 import logging
 import xmlrpclib
+from collections import namedtuple
 from openerp.osv import fields, orm
 from openerp.addons.connector.queue.job import job
+from openerp.addons.connector.connector import ConnectorUnit
 from openerp.addons.connector.exception import MappingError
 from openerp.addons.connector.unit.backend_adapter import BackendAdapter
 from openerp.addons.connector.unit.mapper import (mapping,
@@ -54,8 +56,22 @@ class res_partner(orm.Model):
         'company': fields.char('Company'),
     }
 
+    def copy_data(self, cr, uid, id, default=None, context=None):
+        if default is None:
+            default = {}
+        default['magento_bind_ids'] = False
+        return super(res_partner, self).copy_data(cr, uid, id,
+                                                  default=default,
+                                                  context=context)
 
-# TODO migrate from res.partner (magento fields)
+    def _address_fields(self, cr, uid, context=None):
+        """ Returns the list of address fields that are synced from the parent
+        when the `use_parent_address` flag is set. """
+        fields = super(res_partner, self)._address_fields(cr, uid, context=context)
+        fields.append('company')
+        return fields
+
+
 class magento_res_partner(orm.Model):
     _name = 'magento.res.partner'
     _inherit = 'magento.binding'
@@ -104,6 +120,12 @@ class magento_res_partner(orm.Model):
         'taxvat': fields.char('Magento VAT'),
         'newsletter': fields.boolean('Newsletter'),
         'guest_customer': fields.boolean('Guest Customer'),
+        'consider_as_company': fields.boolean(
+            'Considered as company',
+            help="An account imported with a 'company' in "
+                 "the billing address is considered as a company.\n "
+                 "The partner takes the name of the company and "
+                 "is not merged with the billing address."),
     }
 
     _sql_constraints = [
@@ -252,43 +274,28 @@ class PartnerImport(MagentoImportSynchronizer):
                                                          'magento.res.partner.category')
             importer.run(record['group_id'])
 
+    @property
+    def mapper(self):
+        """ Return an instance of ``Mapper`` for the synchronization.
+
+        The instanciation is delayed because some synchronisations do
+        not need such an unit and the unit may not exist.
+
+        For ``magento.res.partner``, we have a company mapper and
+        a mapper, ensure we find the correct one here.
+
+        :rtype: :py:class:`~.PartnerImportMapper`
+        """
+        if self._mapper is None:
+            get_unit = self.environment.get_connector_unit
+            self._mapper = get_unit(PartnerImportMapper)
+        return self._mapper
+
     def _after_import(self, partner_binding_id):
         """ Import the addresses """
-        addresses_adapter = self.get_connector_unit_for_model(BackendAdapter,
-                                                              'magento.address')
-        mag_address_ids = addresses_adapter.search(
-                {'customer_id': {'eq': self.magento_id}})
-        if not mag_address_ids:
-            return
-        importer = self.get_connector_unit_for_model(MagentoImportSynchronizer,
-                                                     'magento.address')
-        partner_row = self.model.read(self.session.cr,
-                                      self.session.uid,
-                                      partner_binding_id,
-                                      ['openerp_id'],
-                                      context=self.session.context)
-        partner_openerp_id = partner_row['openerp_id'][0]
-        mag_addresses = {} # mag_address_id -> True if address is linked to existing partner,
-                           #                   False otherwise
-        if len(mag_address_ids) == 1:
-            mag_addresses[mag_address_ids[0]] = True
-        else:
-            billing_address = False
-            for address_id in mag_address_ids:
-                magento_record = addresses_adapter.read(address_id)
-
-                if magento_record['is_default_billing']:
-                    mag_addresses[address_id] = True
-                    billing_address = True
-                else:
-                    mag_addresses[address_id] = False
-            if not billing_address:
-                mag_addresses[min(mag_addresses)] = True
-        for address_id, to_link in mag_addresses.iteritems():
-            importer.run(address_id,
-                         partner_binding_id,
-                         partner_openerp_id,
-                         to_link)
+        get_unit = self.get_connector_unit_for_model
+        book = get_unit(PartnerAddressBook, 'magento.address')
+        book.import_addresses(self.magento_id, partner_binding_id)
 
 
 @magento
@@ -361,6 +368,10 @@ class PartnerImportMapper(ImportMapper):
     def customer(self, record):
         return {'customer': True}
 
+    @mapping
+    def type(self, record):
+        return {'type': 'default'}
+
     @only_create
     @mapping
     def openerp_id(self, record):
@@ -376,79 +387,99 @@ class PartnerImportMapper(ImportMapper):
             return {'openerp_id': partner_ids[0]}
 
 
-@magento
-class AddressAdapter(GenericAdapter):
-    _model_name = 'magento.address'
-    _magento_model = 'customer_address'
-
-    def search(self, filters=None):
-        """ Search records according to some criterias
-        and returns a list of ids
-
-        :rtype: list
-        """
-        return [int(row['customer_address_id']) for row
-                in self._call('%s.list' % self._magento_model,
-                              [filters] if filters else [{}])]
+AddressInfos = namedtuple('AddressInfos', ['magento_record',
+                                           'partner_binding_id',
+                                           'merge'])
 
 
 @magento
-class AddressImport(MagentoImportSynchronizer):
-    _model_name = ['magento.address']
+class PartnerAddressBook(ConnectorUnit):
+    """ Import all addresses from the address book of a customer.
 
-    def run(self, magento_id, magento_partner_id, partner_id, link_with_partner):
-        """ Run the synchronization """
-        self.partner_id = partner_id
-        self.magento_partner_id = magento_partner_id
-        self.link_with_partner = link_with_partner
-        super(AddressImport, self).run(magento_id)
+        This class is responsible to define which addresses should
+        be imported and how (merge with the partner or not...).
+        Then, it delegate the import to the appropriate importer.
 
-    def _define_partner_relationship(self, data):
-        """ Link address with partner or parent company. """
-        if self.link_with_partner:
-            # it won't be imported as an address,
-            # but will be linked with the main res.partner
-            data['openerp_id'] = self.partner_id
-        else:
-            data['parent_id'] = self.partner_id
-            partner = self.session.browse('res.partner',
-                                          self.partner_id)
-            data['lang'] = partner.lang
-        data['magento_partner_id'] = self.magento_partner_id
-        return data
+        This is really intricate. The datamodel are different between
+        Magento and OpenERP and we have many uses cases to cover.
 
-    def _create(self, data):
-        data = self._define_partner_relationship(data)
-        return super(AddressImport, self)._create(data)
+        The first thing is that:
+            - we do not import companies and individuals the same manner
+            - we do not know if an account is a company -> we assume that
+              if we found something in the company field of the billing
+              address, the whole account is a company.
 
+        Differences:
+            - Individuals: we merge the billing address with the partner,
+              so we'll end with 1 entity if the customer has 1 address
+            - Companies: we never merge the addresses with the partner,
+              but we use the company name of the billing address as name
+              of the partner. We also copy the address informations from
+              the billing address as default values.
 
-@magento
-class AddressImportMapper(ImportMapper):
+        More information on:
+        https://bugs.launchpad.net/openerp-connector/+bug/1193281
+    """
     _model_name = 'magento.address'
 
-# TODO fields not mapped:
-#   "suffix"=>"a",
-#   "vat_id"=>"12334",
+    def import_addresses(self, magento_partner_id, partner_binding_id):
+        get_unit = self.get_connector_unit_for_model
+        addresses = self._get_address_infos(magento_partner_id,
+                                            partner_binding_id)
+        for address_id, infos in addresses:
+            importer = get_unit(MagentoImportSynchronizer)
+            importer.run(address_id, infos)
 
-    direct = [
-            ('postcode', 'zip'),
-            ('city', 'city'),
-            ('created_at', 'created_at'),
-            ('updated_at', 'updated_at'),
-            ('telephone', 'phone'),
-            ('fax', 'fax'),
-            ('is_default_billing', 'is_default_billing'),
-            ('is_default_shipping', 'is_default_shipping'),
-            ('company', 'company'),
-        ]
+    def _get_address_infos(self, magento_partner_id, partner_binding_id):
+        get_unit = self.get_connector_unit_for_model
+        adapter = get_unit(BackendAdapter)
+        mag_address_ids = adapter.search({'customer_id':
+                                          {'eq': magento_partner_id}})
+        if not mag_address_ids:
+            return
+        for address_id in mag_address_ids:
+            magento_record = adapter.read(address_id)
 
-    @mapping
-    def names(self, record):
-        # TODO create a glue module for base_surname
-        parts = [part for part in (record['firstname'],
-                                   record.get('middlename'),
-                                   record['lastname']) if part]
-        return {'name': ' '.join(parts)}
+            # defines if the billing address is merged with the partner
+            # or imported as a standalone contact
+            merge = False
+            if magento_record.get('is_default_billing'):
+                if magento_record.get('company'):
+                    # when a company is there, we never merge the contact
+                    # with the partner.
+                    # Copy the billing address on the company
+                    # and use the name of the company for the name
+                    company_mapper = get_unit(CompanyImportMapper,
+                                              'magento.res.partner')
+                    company_mapper.convert(magento_record)
+                    self.session.write('magento.res.partner',
+                                       partner_binding_id,
+                                       company_mapper.data)
+                else:
+                    # for B2C individual customers, merge with the main
+                    # partner
+                    merge = True
+                    # in the case if the billing address no longer
+                    # has a company, reset the flag
+                    self.session.write('magento.res.partner',
+                                       partner_binding_id,
+                                       {'consider_as_company': False})
+            address_infos = AddressInfos(magento_record=magento_record,
+                                         partner_binding_id=partner_binding_id,
+                                         merge=merge)
+            yield address_id, address_infos
+
+
+class BaseAddressImportMapper(ImportMapper):
+    """ Defines the base mappings for the imports
+    in ``res.partner`` (state, country, ...)
+    """
+    direct = [('postcode', 'zip'),
+              ('city', 'city'),
+              ('telephone', 'phone'),
+              ('fax', 'fax'),
+              ('company', 'company'),
+              ]
 
     @mapping
     def state(self, record):
@@ -481,23 +512,6 @@ class AddressImportMapper(ImportMapper):
         return result
 
     @mapping
-    def backend_id(self, record):
-        return {'backend_id': self.backend_record.id}
-
-    @mapping
-    def use_parent_address(self, record):
-        return {'use_parent_address': False}
-
-    @mapping
-    def type(self, record):
-        #TODO select type from sale order datas
-        if record.get('is_default_shipping'):
-            address_type = 'delivery'
-        else:
-            address_type = 'default'
-        return {'type': address_type}
-
-    @mapping
     def title(self, record):
         prefix = record['prefix']
         title_id = False
@@ -511,8 +525,138 @@ class AddressImportMapper(ImportMapper):
                 title_id = self.session.create('res.partner.title',
                                                {'domain': 'contact',
                                                'shortcut': prefix,
-                                               'name' : prefix})
+                                               'name': prefix})
         return {'title': title_id}
+
+
+@magento
+class CompanyImportMapper(BaseAddressImportMapper):
+    """ Special mapping used when we import a company.
+    A company is considered as such when the billing address
+    of an account has something in the 'company' field.
+
+    This is a very special mapping not used in the same way
+    than the other.
+
+    The billing address will exist as a contact,
+    but we want to *copy* the data on the company.
+
+    The input record is the billing address.
+    The mapper returns data which will be written on the
+    main partner, in other words, the company.
+
+    The ``@only_create`` decorator would not have any
+    effect here because the mapper is always called
+    for updates.
+    """
+    _model_name = 'magento.res.partner'
+
+    direct = BaseAddressImportMapper.direct + [
+        ('company', 'name'),
+    ]
+
+    @mapping
+    def consider_as_company(self, record):
+        return {'consider_as_company': True}
+
+
+@magento
+class AddressAdapter(GenericAdapter):
+    _model_name = 'magento.address'
+    _magento_model = 'customer_address'
+
+    def search(self, filters=None):
+        """ Search records according to some criterias
+        and returns a list of ids
+
+        :rtype: list
+        """
+        return [int(row['customer_address_id']) for row
+                in self._call('%s.list' % self._magento_model,
+                              [filters] if filters else [{}])]
+
+
+@magento
+class AddressImport(MagentoImportSynchronizer):
+    _model_name = ['magento.address']
+
+    def run(self, magento_id, address_infos):
+        """ Run the synchronization """
+        self.address_infos = address_infos
+        return super(AddressImport, self).run(magento_id)
+
+    def _get_magento_data(self):
+        """ Return the raw Magento data for ``self.magento_id`` """
+        # we already read the data from the Partner Importer
+        if self.address_infos.magento_record:
+            return self.address_infos.magento_record
+        else:
+            return super(AddressImport, self)._get_magento_data()
+
+    def _define_partner_relationship(self, data):
+        """ Link address with partner or parent company. """
+        partner_binding_id = self.address_infos.partner_binding_id
+        partner_id = self.session.read('magento.res.partner',
+                                       partner_binding_id,
+                                       ['openerp_id'])['openerp_id'][0]
+        if self.address_infos.merge:
+            # it won't be imported as an independant address,
+            # but will be linked with the main res.partner
+            data['openerp_id'] = partner_id
+            data['type'] = 'default'
+        else:
+            data['parent_id'] = partner_id
+            partner = self.session.browse('res.partner', partner_id)
+            data['lang'] = partner.lang
+        data['magento_partner_id'] = self.address_infos.partner_binding_id
+        return data
+
+    def _create(self, data):
+        data = self._define_partner_relationship(data)
+        return super(AddressImport, self)._create(data)
+
+
+@magento
+class AddressImportMapper(BaseAddressImportMapper):
+    _model_name = 'magento.address'
+
+# TODO fields not mapped:
+#   "suffix"=>"a",
+#   "vat_id"=>"12334",
+
+    direct = BaseAddressImportMapper.direct + [
+        ('created_at', 'created_at'),
+        ('updated_at', 'updated_at'),
+        ('is_default_billing', 'is_default_billing'),
+        ('is_default_shipping', 'is_default_shipping'),
+        ('company', 'company'),
+    ]
+
+    @mapping
+    def names(self, record):
+        # TODO create a glue module for base_surname
+        parts = [part for part in (record['firstname'],
+                                   record.get('middlename'),
+                                   record['lastname']) if part]
+        return {'name': ' '.join(parts)}
+
+    @mapping
+    def backend_id(self, record):
+        return {'backend_id': self.backend_record.id}
+
+    @mapping
+    def use_parent_address(self, record):
+        return {'use_parent_address': False}
+
+    @mapping
+    def type(self, record):
+        if record.get('is_default_billing'):
+            address_type = 'invoice'
+        elif record.get('is_default_shipping'):
+            address_type = 'delivery'
+        else:
+            address_type = 'contact'
+        return {'type': address_type}
 
 
 @job